test: extract AgentMCPAction factory for blocked action tests
